requirements.txt
python-dotenv>=1.0.1
openai>=1.40.0
requests>=2.32.3
praw>=7.7.1
APScheduler>=3.10.4
pytz>=2024.1
tweepy>=4.14.0
pymongo>=4.8.0
